Add files via upload
hey guys, heres my version of the ethnocentrism model. I added comments about the parts that i added on to the original one, usually something slong the lines of "GABRIEL ADDED THIS" or "THIS IS NEW" or something. i hope this is the intended way of uploading files otherwise let me know haha